Remove consistency policy
